stage 6 완료

+++ Web Quiz Engine (Kotlin)/task/src/engine/repository/QuizRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface QuizRepository : JpaRepository<Quiz, Long> {
+interface QuizRepository : JpaRepository<Quiz, Int> {
     fun findQuizById(id: Int): Quiz?
 }
